+++ pkg/analysis_server/test/src/services/refactoring/move_top_level_to_file_test.dart
@@ -50,6 +50,8 @@ class MoveTopLevelToFileTest extends RefactoringTest {
 
   Future<void> test_class() async {
     var originalSource = '''
+// File header.
+
 class A {}
 
 class ClassToMove^ {}
@@ -57,6 +59,8 @@ class ClassToMove^ {}
 class B {}
 ''';
     var modifiedSource = '''
+// File header.
+
 class A {}
 
 class B {}
@@ -64,6 +68,8 @@ class B {}
     var declarationName = 'ClassToMove';
     var newFileName = 'class_to_move.dart';
     var newFileContent = '''
+// File header.
+
 class ClassToMove {}
 ''';
     await _singleDeclaration(