+++ pkg/analyzer/lib/src/dart/element/inheritance_manager3.dart
@@ -907,10 +907,7 @@ class InheritanceManager3 {
 
   static bool _isDeclaredInObject(ExecutableElement element) {
     var enclosing = element.enclosingElement;
-    // TODO(scheglov) `is! MixinElement` after the separation.
-    return enclosing is InterfaceElement &&
-        enclosing.supertype == null &&
-        enclosing is! MixinElement;
+    return enclosing is ClassElement && enclosing.isDartCoreObject;
   }
 }